DEV Use abstract methods for base classes
