+++ serpentTools/objects/__init__.py
@@ -1,13 +1,6 @@
 """Objects used to support the parsing."""
 
-
-class NamedObject(object):
-    """Class for named objects like materials and detectors."""
-
-    def __init__(self, name):
-        self.name = name
-
-    def __str__(self):
-        return '<{} {}>'.format(self.__class__.__name__, self.name)
+from serpentTools.objects.containers import (
+        HomogUniv, Detector, BranchContainer)
