+++ serpentTools/objects/base.py
@@ -1,7 +1,15 @@
+"""
+Collection of base classes from which other objects inherit.
+"""
+from abc import ABCMeta, abstractmethod
+
+from six import add_metaclass
+
 from serpentTools.settings import rc
 from serpentTools.messages import info
 
-class BaseReader:
+@add_metaclass(ABCMeta)
+class BaseReader(object):
     """Parent class from which all parsers will inherit.
 
     Parameters
@@ -35,6 +43,7 @@ class BaseReader:
         info("  - done")
         self._postcheck()
 
+    @abstractmethod
     def _read(self):
         """Read the file and store the data.
 
@@ -43,20 +52,32 @@ class BaseReader:
             This read function has not been implemented yet
 
         """
-        raise NotImplementedError
+        pass
 
+    @abstractmethod
     def _precheck(self):
         """Pre-checking, e.g., make sure Serpent did not
         exit abnormally, or disk ran out of space while parsing.
         """
         pass
 
+    @abstractmethod
     def _postcheck(self):
         """Make sure data looks reasonable. Could possibly check for
         negative cross sections, negative material densitites, etc (which
         can happen if using the reprocessing interface incorrectly).
         """
         pass
+
+
+class NamedObject(object):
+    """Class for named objects like materials and detectors."""
+
+    def __init__(self, name):
+        self.name = name
+
+    def __str__(self):
+        return '<{} {}>'.format(self.__class__.__name__, self.name)
 
 
 class MaterialReader(BaseReader):
