+++ serpentTools/objects/materials.py
@@ -5,7 +5,7 @@ from matplotlib import pyplot
 
 from serpentTools.messages import warning, debug
 from serpentTools.plot import magicPlotDocDecorator, formatPlot, DEPLETION_PLOT_LABELS
-from serpentTools.objects import NamedObject
+from serpentTools.objects.base import NamedObject
 from serpentTools.utils import convertVariableName
